changed field type

+++ playground/django_test/osm_implementation/random_walker_engine/models.py
@@ -18,8 +18,8 @@ class User(models.Model):
 
 class Location(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    origin_lat = models.FloatField()
-    origin_lng = models.FloatField()
-    destination_lat = models.FloatField()
-    destination_lng = models.FloatField()
+    origin_lat = models.DecimalField(max_digits=20, decimal_places=10)
+    origin_lng = models.DecimalField(max_digits=20, decimal_places=10)
+    destination_lat = models.DecimalField(max_digits=20, decimal_places=10)
+    destination_lng = models.DecimalField(max_digits=20, decimal_places=10)
     date_generation = models.DateTimeField()
